dom/url/tests/test_url_empty_port.html


<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=930450
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 930450</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=930450">Mozilla Bug 930450</a>
<p id="display"></p>
<div id="content" style="display: none">
  <iframe name="x" id="x"></iframe>
  <iframe name="y" id="y"></iframe>
</div>
<pre id="test">
</pre>
  <a id="link" href="http://www.example.com:8080">foobar</a>
  <area id="area" href="http://www.example.com:8080" />
  <script type="application/javascript">

  var url = new URL('http://www.example.com:8080');
  is(url.port, '8080', 'URL.port is 8080');
  url.port = '';
  is(url.port, '', 'URL.port is \'\'');
  url.port = 0;
  is(url.port, '0', 'URL.port is 0');

  var link = document.getElementById("link");
  is(link.port, '8080', 'URL.port is 8080');
  link.port = '';
  is(link.href, 'http://www.example.com/', "link.href matches");
  is(link.port, '', 'URL.port is \'\'');
  link.port = 0;
  is(link.href, 'http://www.example.com:0/', "link.href matches");
  is(link.port, '0', 'URL.port is 0');

  var area = document.getElementById("area");
  is(area.port, '8080', 'URL.port is 8080');
  area.port = '';
  is(area.href, 'http://www.example.com/', "area.href matches");
  is(area.port, '', 'URL.port is \'\'');
  area.port = 0;
  is(area.href, 'http://www.example.com:0/', "area.href matches");
  is(area.port, '0', 'URL.port is 0');

  </script>
</body>
</html>
